Add country details
