+++ src/index.js
@@ -6,6 +6,10 @@ import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
 import store from './redux/configureStore';
+import date from './components/Date';
+import { getAllCountriesData } from './redux/countries';
+
+store.dispatch(getAllCountriesData(date));
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
