+++ src/pages/DetailsPage.js
@@ -1,8 +1,10 @@
 import React from 'react';
+import Regions from '../components/Regions';
 
 const DetailsPage = () => (
   <div>
     <h3>Details</h3>
+    <Regions />
   </div>
 );
